refactor: apply SOLID principles across the framework
[S] SRP:
- Extract CssRuleFilter, SelectorMatcher, HtmlSelectorExtractor from
  RemoveUnusedCss (Performance/Css/)
- Extract AuditLogRenderer from AuditLogAdminPage (Security/AuditLog/)

[O] OCP:
- Add getBundles() to ExtensionInterface, introduce AbstractExtension
  base class with default implementation wrapping getBundle()
- Migrate all 16 extensions from `implements ExtensionInterface` to
  `extends AbstractExtension`
- Remove method_exists() hack in WordPressContainer::getBundles()
- Add extensible hook runner system to HookRegistry via addRunner()

[L] LSP:
- Fix setSetArgs typo in PostMetaStructureInterface (already corrected)

[I] ISP:
- Segregate PostMetaStructureInterface into 3 focused interfaces:
  MetaKeyAwareInterface, MetaSchemaInterface, MetaAccessControlInterface
- PostMetaStructureInterface now composes all three for backward compat

[D] DIP:
- Extract HookRegistryInterface in Hooks/Contracts/
- WordPressContainer now depends on HookRegistryInterface abstraction
  instead of concrete HookRegistry class
- Register interface alias in HooksExtension
