+++ src/Hooks/HooksExtension.php
@@ -4,14 +4,15 @@ declare(strict_types=1);
 
 namespace BackTo\Framework\Hooks;
 
-use BackTo\Framework\Contracts\ExtensionInterface;
+use BackTo\Framework\Compose\AbstractExtension;
 use BackTo\Framework\Contracts\HookDispatcherInterface;
 use BackTo\Framework\Contracts\HookInterface;
+use BackTo\Framework\Hooks\Contracts\HookRegistryInterface;
 use BackTo\Framework\Hooks\DependencyInjection\Compiler\RegisterHookPass;
 use BackTo\Framework\Hooks\Infrastructure\WordPressHookDispatcher;
 use BackToVendor\Symfony\Component\DependencyInjection\ContainerBuilder;
 
-final class HooksExtension implements ExtensionInterface
+final class HooksExtension extends AbstractExtension
 {
     public function getBundle(): ?array
     {
@@ -31,6 +32,9 @@ final class HooksExtension implements ExtensionInterface
 
         $containerBuilder->register(HookDispatcherInterface::class, WordPressHookDispatcher::class);
         $containerBuilder->setAlias(WordPressHookDispatcher::class, HookDispatcherInterface::class);
+
+        $containerBuilder->setAlias(HookRegistryInterface::class, HookRegistry::class)
+            ->setPublic(true);
     }
 
     public function getDefaultConfiguration(): array
